seps-plugin/scripts/driver/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
psd-tools>=1.9.28
pypdfium2>=4.25.0
